+++ test/e2e/fragmented_send.c
@@ -119,7 +119,7 @@ void onOpen(ws_conn_t *conn) {
 
 void onMsg(ws_conn_t *conn, void *msg, size_t n, uint8_t opcode) {}
 
-void onDisconnect(ws_conn_t *conn, int err) {
+void onDisconnect(ws_conn_t *conn, unsigned long  err) {
   assert(ws_conn_ctx(conn));
   free(ws_conn_ctx(conn));
   ws_conn_set_ctx(conn, NULL);
