+++ test/e2e/user_epoll_test.c
@@ -63,7 +63,7 @@ void onOpen(ws_conn_t *conn) {}
 
 void onMsg(ws_conn_t *conn, void *msg, size_t n, uint8_t opcode) {}
 
-void onDisconnect(ws_conn_t *conn, int err) {}
+void onDisconnect(ws_conn_t *conn, unsigned long  err) {}
 
 int main(void) {
   printf("interval example starting on 9919\n");
